refactor all

+++ app/Http/Controllers/Admin/PostController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Category;
+use App\Tag;
 use App\Post;
+use App\Category;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use App\Http\Controllers\Controller;
@@ -18,6 +19,8 @@ class PostController extends Controller
             'max:100',
         ],
         'category_id' => 'required:integer|exists:categories,id',
+        'tags'          => 'array',
+        'tags.*'        => 'integer|exists:tags,id',
         'title'=>'required|string|max:100',
         'image'=>'string|max:100',
         'uploaded_img'=>'image|max:1024',
@@ -46,9 +49,11 @@ class PostController extends Controller
     public function create()
     {
         $categories = Category::all('id', 'name');
+        $tags       = Tag::all();
 
        return view('admin.posts.create', [
         'categories' => $categories,
+        'tags'          => $tags,
        ]);
     }
 
@@ -78,6 +83,8 @@ class PostController extends Controller
         $post->excerpt        = $data['excerpt'];
         $post->save();
 
+        $post->tags()->attach($data['tags']);
+
         return redirect()->route('admin.posts.show', [
             'post' => $post
         ]);
@@ -104,8 +111,14 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $categories = Category::all('id', 'name');
+        $tags       = Tag::all();
+
+
         return view('admin.posts.edit', [
             'post' => $post,
+            'categories'    => $categories,
+            'tags'          => $tags,
         ]);
     }
 
@@ -124,9 +137,12 @@ class PostController extends Controller
 
         $data = $request->all();
 
-        $img_path = Storage::put('uploads', $data['uploaded_img']);
-
-        Storage::delete($post->uploaded_img);
+        if (isset($data['uploaded_img'])) {
+            $img_path = Storage::put('uploads', $data['uploaded_img']);
+            Storage::delete($post->uploaded_img);
+        } else {
+            $img_path = $post->uploaded_img;
+        }
 
         $post->slug           = $data['slug'];
         $post->title          = $data['title'];
@@ -135,6 +151,8 @@ class PostController extends Controller
         $post->content        = $data['content'];
         $post->excerpt        = $data['excerpt'];
         $post->update();
+
+        $post->tags()->sync($data['tags']);
 
         return redirect()->route('admin.posts.show', [
             'post' => $post
@@ -149,8 +167,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $post->tags()->detach();
+
         $post->delete();
-       return redirect()->route('admin.posts.index', ['post' => $post])->with('success_delete', $post);
+
+        return redirect()->route('admin.posts.index')->with('success_delete', $post);
 
     }
 }
